Most Features
Implemented most upgrades, a basic shop, timer that kills the player, weak spots, player clicking to hurt enemy.

Main thing missing is the number conversion system that prevents crashes.

+++ Assets/Scripts/UpgradeManager.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 
 public class UpgradeManager
 {
-    public Dictionary<string, float> Upgrades = new Dictionary<string, float>()
+    public Dictionary<string, int> Upgrades = new Dictionary<string, int>()
     {
         //Guide to Keys: { Upgrade Name, Upgrade Tier }
